Re-work weblayer's permission tests to work with androidx roll
These tests started failing on the androidx autoroller (see bug).

The main issue revealed by the roll, is that the Context given by
fragments need to have a .getApplicationContext() that returns an
Application (or ContextWrapper around Application). The current test
used InMemorySharedPreferencesContext for the base context of the
Activity, which just returns "this" from getApplicationContext().

WebLayer passes the result of getApplicationContext() to the WebLayer's
ClassLoader / side of the world, and so in order to override
checkPermission() on the Context object being used by WebLayer, the test
now overrides the Application context via ContextUtils.

It turns out that the fragment also queries the attached activity's
checkPermission(), and so that needs to be overridden explicitly as
well. I thus made InstrumtationActivity's checkPermission() delegate to the
application context.

Bug: 1288160
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3401706
Commit-Queue: Andrew Grieve <agrieve@chromium.org>
Cr-Commit-Position: refs/heads/main@{#961528}

+++ weblayer/browser/android/javatests/src/org/chromium/weblayer/test/InputTypesTest.java
@@ -8,7 +8,6 @@ import android.Manifest;
 import android.annotation.TargetApi;
 import android.app.Activity;
 import android.content.ClipData;
-import android.content.Context;
 import android.content.Intent;
 import android.content.pm.PackageManager;
 import android.net.Uri;
@@ -29,11 +28,11 @@ import org.junit.Rule;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 
-import org.chromium.base.Function;
+import org.chromium.base.ContextUtils;
+import org.chromium.base.test.util.ApplicationContextWrapper;
 import org.chromium.base.test.util.CallbackHelper;
 import org.chromium.base.test.util.Criteria;
 import org.chromium.base.test.util.CriteriaHelper;
-import org.chromium.base.test.util.InMemorySharedPreferencesContext;
 import org.chromium.base.test.util.MinAndroidSdkLevel;
 import org.chromium.base.test.util.UrlUtils;
 import org.chromium.content_public.browser.test.util.TestThreadUtils;
@@ -129,43 +128,45 @@ public class InputTypesTest {
 
     @Before
     public void setUp() throws Exception {
-        Bundle extras = new Bundle();
-        // We need to override the context with which to create WebLayer.
-        extras.putBoolean(InstrumentationActivity.EXTRA_CREATE_WEBLAYER, false);
-
-        Function<Context, Context> applicationContextBuilder = (baseContext) -> {
-            return new InMemorySharedPreferencesContext(baseContext) {
-                @Override
-                public int checkPermission(String permission, int pid, int uid) {
-                    if (permission.equals(Manifest.permission.CAMERA)) {
-                        return mCameraPermission;
+        ContextUtils.initApplicationContextForTests(
+                new ApplicationContextWrapper(ContextUtils.getApplicationContext()) {
+                    @Override
+                    public int checkPermission(String permission, int pid, int uid) {
+                        if (permission.equals(Manifest.permission.CAMERA)) {
+                            return mCameraPermission;
+                        }
+                        return super.checkPermission(permission, pid, uid);
                     }
-                    return getBaseContext().checkPermission(permission, pid, uid);
-                }
-            };
-        };
-        InstrumentationActivity.setActivityContextBuilder(applicationContextBuilder);
+                });
+        ActivityCompat.setPermissionCompatDelegate(mPermissionCompatDelegate);
 
-        InstrumentationActivity activity = mActivityTestRule.launchShell(extras);
-        TestThreadUtils.runOnUiThreadBlocking(
-                () -> { activity.loadWebLayerSync(activity.getApplicationContext()); });
-        mActivityTestRule.navigateAndWait(mActivityTestRule.getTestDataURL("input_types.html"));
         mTestDir = new File(UrlUtils.getIsolatedTestFilePath("weblayer"));
         if (!mTestDir.exists()) mTestDir.mkdir();
         mTempFile = File.createTempFile("file", null, mTestDir);
-        activity.setIntentInterceptor(mIntentInterceptor);
-        ActivityCompat.setPermissionCompatDelegate(mPermissionCompatDelegate);
-
         Intent response = new Intent();
         response.setData(Uri.fromFile(mTempFile));
         mIntentInterceptor.setResponse(Activity.RESULT_OK, response);
+
+        Bundle extras = new Bundle();
+        // We need to override the context with which to create WebLayer.
+        extras.putBoolean(InstrumentationActivity.EXTRA_CREATE_WEBLAYER, false);
+        InstrumentationActivity activity = mActivityTestRule.launchShell(extras);
+        activity.setIntentInterceptor(mIntentInterceptor);
+
+        TestThreadUtils.runOnUiThreadBlocking(
+                () -> activity.loadWebLayerSync(ContextUtils.getApplicationContext()));
+        mActivityTestRule.navigateAndWait(mActivityTestRule.getTestDataURL("input_types.html"));
     }
 
     @After
     public void tearDown() {
-        mTempFile.delete();
-        mTestDir.delete();
         ActivityCompat.setPermissionCompatDelegate(null);
+        if (mTempFile != null) {
+            mTempFile.delete();
+        }
+        if (mTestDir != null) {
+            mTestDir.delete();
+        }
     }
 
     @Test
